+++ package.json
@@ -1,6 +1,6 @@
 {
   "name": "experiment-comparator",
-  "version": "1.0.1",
+  "version": "1.0.2",
   "description": "A modern desktop application for comparing machine learning model experiment results",
   "type": "module",
   "scripts": {
