+++ src/store/index.ts
@@ -1,0 +1,8 @@
+// 导出主要的store
+export { useAppStore } from './appStore';
+
+// 导出所有selector hooks
+export * from './selectors';
+
+// 导出类型定义
+export * from './types';
